Fixes theme view bug
